唉, 拆分
