+++ src/callbacks/chatmember.py
@@ -1,9 +1,11 @@
 import asyncio
+
 from telegram import Chat, ChatMember, ChatMemberUpdated, Update
 from telegram.ext import ContextTypes
 
+from ..dao.association import delete_association_in_chat
 from ..logger import logger
-from ..database import dao
+from ..service.chat import delete_chat_data
 
 
 def extract_status_change(chat_member_update: ChatMemberUpdated):
@@ -55,7 +57,7 @@ async def track_chats(update: Update, context: ContextTypes.DEFAULT_TYPE) -> Non
 
         elif was_member and not is_member:
             logger.debug(f"{cause_name} 将bot移出群组 {chat.title}")
-            dao.delete_chat_data(chat)
+            delete_chat_data(chat)
 
     elif not was_member and is_member:
         logger.debug(f"{cause_name} 将bot添加到频道 {chat.title}")
@@ -72,7 +74,7 @@ async def track_chats(update: Update, context: ContextTypes.DEFAULT_TYPE) -> Non
 async def on_member_left(update: Update, context: ContextTypes.DEFAULT_TYPE):
     left_user = update.effective_message.left_chat_member
     logger.debug(f"{left_user.full_name} 退出了群聊 {update.effective_chat.title}")
-    dao.remove_user_from_chat(left_user, update.effective_chat)
+    delete_association_in_chat(update.effective_chat, left_user.id)
 
 
 async def on_member_join(update: Update, context: ContextTypes.DEFAULT_TYPE):
