+++ src/callbacks/start.py
@@ -3,7 +3,12 @@ from telegram.ext import ContextTypes
 
 from ..logger import logger
 from ..common.message import message_recorder
-from ..database import dao
+
+from ..dao.db import db
+from ..dao.user import (
+    add_user,
+    get_user_by_id,
+)
 from ..common.user import download_big_avatar
 
 _start_bot_markup = InlineKeyboardMarkup(
@@ -45,11 +50,11 @@ async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
             text="Nya~",
             reply_markup=_start_bot_markup,
         )
-    db_bot_user = dao.get_user_by_id(context.bot.id)
+    db_bot_user = get_user_by_id(context.bot.id)
     if not db_bot_user:
-        db_bot_user = dao.add_user((await context.bot.get_me()))
+        db_bot_user = add_user((await context.bot.get_me()))
         db_bot_user.avatar_big_blob = await download_big_avatar(context.bot.id, context)
-        dao.commit()
+        db.commit()
     photo = db_bot_user.avatar_big_id or db_bot_user.avatar_big_blob
     sent_message = await update.effective_message.reply_photo(
         photo=photo,
@@ -57,7 +62,7 @@ async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
         reply_markup=_start_bot_markup,
     )
     db_bot_user.avatar_big_id = sent_message.photo[-1].file_id
-    dao.commit()
+    db.commit()
 
 
 async def _start_in_group(update: Update, context: ContextTypes.DEFAULT_TYPE):
