+++ src/callbacks/userdata.py
@@ -1,14 +1,24 @@
-from telegram import InlineKeyboardButton, InlineKeyboardMarkup, Update, InputMediaPhoto
-from telegram.ext import ContextTypes
-from ..database import dao
-
-from ..logger import logger
-from ..common.user import get_big_avatar_bytes, get_small_avatar_bytes, get_user_info
-from ..common.waifu import get_user_waifu_info
-from ..common.utils import back_home_markup
-from .jobs import reset_user_cd
 from math import ceil
+
+from telegram import InlineKeyboardButton, InlineKeyboardMarkup, InputMediaPhoto, Update
+from telegram.ext import ContextTypes
 from telegram.helpers import escape_markdown
+
+from ..common.user import get_big_avatar_bytes, get_small_avatar_bytes, get_user_info
+from ..common.utils import back_home_markup
+from ..common.waifu import get_user_waifu_info
+from ..dao.db import db
+from ..dao.quote import delete_quote_by_link
+from ..dao.user import (
+    get_user_by_id,
+    get_user_quotes_count,
+    get_user_quotes_page,
+)
+from ..logger import logger
+from ..service.waifu import (
+    refresh_user_all_waifu,
+)
+from .jobs import reset_user_cd
 
 _user_data_manage_markup = InlineKeyboardMarkup(
     [
@@ -24,7 +34,7 @@ async def user_data_manage(update: Update, context: ContextTypes.DEFAULT_TYPE):
     user = update.effective_user
     query = update.callback_query
     logger.info(f"({user.name}) <user data manage>")
-    db_user = dao.get_user_by_id(user.id)
+    db_user = get_user_by_id(user.id)
     info = get_user_info(user)
     if db_user.avatar_big_id:
         await query.edit_message_media(
@@ -77,13 +87,13 @@ async def user_data_refresh(update: Update, context: ContextTypes.DEFAULT_TYPE):
         )
         avatar_big_id = sent_message.photo[-1].file_id
         await sent_message.delete()
-    db_user = dao.get_user_by_id(user.id)
+    db_user = get_user_by_id(user.id)
     db_user.username = username
     db_user.full_name = full_name
     db_user.avatar_big_blob = avatar_big_blog
     db_user.avatar_small_blob = avatar_small_blog
     db_user.avatar_big_id = avatar_big_id
-    dao.commit()
+    db.commit()
     info = get_user_info(user)
     if avatar_big_id:
         await query.edit_message_media(
@@ -106,7 +116,7 @@ async def user_waifu_manage(update: Update, context: ContextTypes.DEFAULT_TYPE):
     if "divorce" in query_data:
         await _divorce_ask(update, context)
         return
-    db_user = dao.get_user_by_id(update.effective_user.id)
+    db_user = get_user_by_id(update.effective_user.id)
     if "set_waifu_mention" in query_data:
         db_user.waifu_mention = not db_user.waifu_mention
     set_mention_text = "别@你" if db_user.waifu_mention else "抽到你时@你"
@@ -123,7 +133,7 @@ async def user_waifu_manage(update: Update, context: ContextTypes.DEFAULT_TYPE):
     )
     text = get_user_waifu_info(update.effective_user)
     await query.edit_message_caption(caption=text, reply_markup=waifu_manage_markup)
-    dao.commit()
+    db.commit()
 
 
 _divorce_ask_markup = InlineKeyboardMarkup(
@@ -141,11 +151,11 @@ async def _divorce_ask(update: Update, context: ContextTypes.DEFAULT_TYPE):
     if "divorce_confirm" in query.data:
         await _divorce_confirm(update, context)
         return
-    db_user = dao.get_user_by_id(update.effective_user.id)
+    db_user = get_user_by_id(update.effective_user.id)
     if not db_user.is_married:
         await query.answer("可是...你还没有结婚呀qwq", show_alert=True, cache_time=15)
         return
-    married_waifu = dao.get_user_by_id(db_user.married_waifu_id)
+    married_waifu = get_user_by_id(db_user.married_waifu_id)
 
     if query.message.photo and married_waifu.avatar_big_id:
         await query.edit_message_media(
@@ -163,16 +173,16 @@ async def _divorce_ask(update: Update, context: ContextTypes.DEFAULT_TYPE):
 
 
 async def _divorce_confirm(update: Update, context: ContextTypes.DEFAULT_TYPE):
-    db_user = dao.get_user_by_id(update.effective_user.id)
+    db_user = get_user_by_id(update.effective_user.id)
     query = update.callback_query
-    married_waifu = dao.get_user_by_id(db_user.married_waifu_id)
+    married_waifu = get_user_by_id(db_user.married_waifu_id)
     db_user.is_married = False
     db_user.married_waifu_id = None
     married_waifu.is_married = False
     married_waifu.married_waifu_id = None
-    dao.commit()
-    dao.refresh_user_all_waifu(db_user)
-    dao.refresh_user_all_waifu(married_waifu)
+    db.commit()
+    refresh_user_all_waifu(db_user)
+    refresh_user_all_waifu(married_waifu)
     await query.edit_message_caption(
         caption="_愿你有一天和重要之人重逢_",
         parse_mode="MarkdownV2",
@@ -189,7 +199,7 @@ async def delete_user_quote(update: Update, context: ContextTypes.DEFAULT_TYPE):
         await _user_quote_manage(update, context)
         return
     quote_link = query.data.split(" ")[1]
-    dao.delete_quote_by_link(quote_link)
+    delete_quote_by_link(quote_link)
     await query.answer("已删除", show_alert=False, cache_time=5)
     await _user_quote_manage(update, context)
 
@@ -200,7 +210,7 @@ async def _user_quote_manage(update: Update, context: ContextTypes.DEFAULT_TYPE)
     logger.info(f"({user.name}) <user quote manage>")
     page = int(query.data.split(" ")[-1]) if len(query.data.split(" ")) > 1 else 1
     page_size = 5
-    quotes_count = dao.get_user_quotes_count(user)
+    quotes_count = get_user_quotes_count(user)
     max_page = ceil(quotes_count / page_size)
     if quotes_count == 0:
         caption = (
@@ -216,7 +226,7 @@ async def _user_quote_manage(update: Update, context: ContextTypes.DEFAULT_TYPE)
         return
     text = f"你的语录: 共{quotes_count}条; 第{page}/{max_page}页\n"
     text += "点击序号删除语录\n\n"
-    quotes = dao.get_user_quotes_page(user, page, page_size)
+    quotes = get_user_quotes_page(user, page, page_size)
     keyboard, line = [], []
     for index, quote in enumerate(quotes):
         quote_content = (
