+++ src/common/waifu.py
@@ -2,9 +2,15 @@ from typing import Generator
 
 from telegram import Chat, User
 
-from ..database import dao
-from ..database.model import ChatData, UserData
+from ..dao.chat import get_chat_members
+from ..dao.user import add_user
 from ..logger import logger
+from ..models.models import ChatData, UserData
+from ..service.waifu import (
+    get_user_waifu_in_chat_exclude_married,
+    get_user_waifus_of_with_chat,
+    get_user_waifus_with_chat,
+)
 
 
 def get_chat_waifu_relationships(
@@ -12,9 +18,9 @@ def get_chat_waifu_relationships(
 ) -> Generator[tuple[int, int], None, None]:
     # relationships: a generator that yields (int, int) for (user_id, waifu_id)
     logger.debug(f"Get chat waifu relationships for {chat.title}<{chat.id}>")
-    members = dao.get_chat_members(chat)
+    members = get_chat_members(chat)
     for member in members:
-        waifu = dao.get_user_waifu_in_chat_exclude_married(member, chat)
+        waifu = get_user_waifu_in_chat_exclude_married(member, chat)
         if waifu:
             yield (member.id, waifu.id)
 
@@ -32,13 +38,13 @@ def get_chat_waifu_info_dict(
 
 def get_user_waifu_info(user: User | UserData) -> str:
     logger.debug(f"Get user waifu info for {user.full_name}<{user.id}>")
-    db_user = dao.add_user(user)
+    db_user = add_user(user)
     text = f"""
 是否@你: {db_user.waifu_mention}
 已婚: {db_user.is_married}
 已婚老婆id: {db_user.married_waifu_id}
 """
-    waifus_with_chat = dao.get_user_waifus_with_chat(user)
+    waifus_with_chat = get_user_waifus_with_chat(user)
     if not waifus_with_chat:
         text += "\n你今天还没有老婆哦"
     else:
@@ -47,7 +53,7 @@ def get_user_waifu_info(user: User | UserData) -> str:
         text += "\n你今天的老婆们(最多显示233条):\n"
         for waifu, chat in waifus_with_chat:
             text += f"{waifu.full_name} ({chat.title})\n"
-    waifus_of_with_chat = dao.get_user_waifus_of_with_chat(user)
+    waifus_of_with_chat = get_user_waifus_of_with_chat(user)
     if not waifus_of_with_chat:
         text += "\n今天还没有人把你当老婆哦"
     else:
